# WebEcomerceStoreAPI/WebEcomerceStoreAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebEcomerce API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Categories/Pagination": {
      "get": {
        "tags": [
          "Categories"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cursorId",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "format": "int32",
            "default": 5
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/IBussinessResult"
            }
          }
        }
      }
    },
    "/api/Categories/GetName": {
      "get": {
        "tags": [
          "Categories"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categoryName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/IBussinessResult"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "IBussinessResult": {
      "type": "object",
      "properties": {
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "data": { }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}